Coupling of Endpoints for readability: - Moved from Program to own dedicated file - Moved lambda functions to own defined functions - Added endpoint group
